python/projetos/django/api_restful_escola/requirements.txt
asgiref==3.8.1
Django==5.0.3
djangorestframework==3.15.2
Markdown==3.6
sqlparse==0.5.0
tzdata==2024.1
